better refactoring
