+++ swiftwinrt/code_writers/class_writers.cpp
@@ -1,6 +1,7 @@
 #include "writer_helpers.h"
 #include "class_writers.h"
 #include "utility/type_helpers.h"
+#include "utility/swift_codegen_utils.h"
 namespace swiftwinrt
 {
     void write_class(writer& w, class_type const& type)
