+++ swiftwinrt/utility/metadata_helpers.h
@@ -148,42 +148,13 @@ namespace swiftwinrt
     const class_type* try_get_exclusive_to(writer& w, typedef_base const& type);
     std::optional<attributed_type> try_get_factory_info(writer& w, typedef_base const& type);
     std::tuple<MethodDef, MethodDef> get_property_methods(Property const& prop);
-    std::string get_swift_name(interface_info const& iface);
-    std::string put_in_backticks_if_needed(std::string name);
-    std::string to_camel_case(std::string_view name);
-    std::string get_swift_name(MethodDef const& method);
-    std::string get_swift_member_name(std::string_view name);
-    std::string get_swift_name(Property const& property);
-    std::string get_swift_name(Event const& event);
-    std::string get_swift_name(Field const& field);
-    std::string get_swift_name(Param const& param);
-    std::string local_swift_param_name(std::string const& param_name);
-    std::string local_swift_param_name(std::string_view param_name);
-    std::string local_swift_param_name(function_param const& param);
-    std::string get_swift_name(function_param const& param);
-    std::string get_swift_name(function_return_type const& return_type);
-    std::string get_swift_name(property_def const& property);
-    std::string get_swift_name(function_def const& function);
-    std::string get_swift_name(struct_member const& member);
-    std::string_view get_abi_name(struct_member const& member);
-    std::string_view remove_backtick(std::string_view name);
-    std::string internal_namepace(std::string prefix, std::string_view ns);
-    std::string abi_namespace(std::string_view ns);
-    std::string impl_namespace(std::string_view ns);
-    std::string abi_namespace(TypeDef const& type);
-    std::string abi_namespace(metadata_type const& type);
-    std::string abi_namespace(metadata_type const* type);
     winmd::reader::ElementType underlying_enum_type(winmd::reader::TypeDef const& type);
     std::array<char, 37> type_iid(winmd::reader::TypeDef const& type);
     param_category get_category(metadata_type const* type, TypeDef* signature_type = nullptr);
     std::string_view get_full_type_name(generic_inst const& type);
     bool is_struct_blittable(struct_type const& type);
     bool is_struct(metadata_type const& type);
-    bool needs_wrapper(param_category category);
     bool is_overridable(metadata_type const& type);
-
-    std::string_view get_abi_name(winmd::reader::MethodDef const& method);
-    std::string_view get_abi_name(function_def const& method);
 
     std::string_view get_name(winmd::reader::MethodDef const& method);
     std::string_view get_name(function_def const& method);
